Integrated patch from woods@most.weird.com (Greg A. Woods)
Subject: [cricket-users] update of Livingston Portmaster lines monitoring....

The portmaster monitoring script included in the distribution didn't do
very much for me with PortMaster-3s running ComOS 3.9b27.  I did a few
quick hacks to get it to at least work but as I note in some new
comments there are a number of other things that could probably be done
better too.

Note that I've moved the interface mib stuff up to the top-level
Defaults file and thus I have views for the ethernet interface on the
portmasters too.

Also note that this probably breaks any support for PM-2s so I deleted
mention of them for now....